Factory Method: decouple traits and implementations
The aim is to demonstrate a value of the Factory Method
by decoupling interfaces and implementations.
Although, the full decoupling is going to be demonstrated in
the Abstract Factory example.
