--- creational/factory-method/render-dialog/button/mod.rs
@@ -1,7 +1,0 @@
-pub mod html;
-pub mod windows;
-
-pub trait Button {
-    fn render(&self);
-    fn on_click(&self);
-}
