--- creational/factory-method/render-dialog/dialog/html.rs
@@ -1,11 +1,0 @@
-use super::Dialog;
-
-use crate::button::{html::HtmlButton, Button};
-
-pub struct HtmlDialog;
-
-impl Dialog for HtmlDialog {
-    fn create_button(&self) -> Box<dyn Button> {
-        Box::new(HtmlButton)
-    }
-}
