+++ creational/factory-method/render-dialog/gui.rs
@@ -1,12 +1,13 @@
-pub mod html;
-pub mod windows;
-
-use crate::button::Button;
+pub trait Button {
+    fn render(&self);
+    fn on_click(&self);
+}
 
 /// Dialog has a factory method `create_button`.
 ///
 /// It creates different buttons depending on a factory implementation.
 pub trait Dialog {
+    /// The factory method. It must be overridden with a concrete implementation.
     fn create_button(&self) -> Box<dyn Button>;
 
     fn render(&self) {
