+++ creational/factory-method/render-dialog/init.rs
@@ -1,0 +1,14 @@
+use crate::gui::Dialog;
+use crate::html_gui::HtmlDialog;
+use crate::windows_gui::WindowsDialog;
+
+pub fn initialize() -> &'static dyn Dialog {
+    // The dialog type is selected depending on the environment settings or configuration.
+    if cfg!(windows) {
+        println!("-- Windows detected, creating Windows GUI --");
+        return &WindowsDialog;
+    } else {
+        println!("-- No OS detected, creating the HTML GUI --");
+        return &HtmlDialog;
+    };
+}
