+++ creational/factory-method/render-dialog/windows_gui.rs
@@ -1,4 +1,4 @@
-use super::Button;
+use crate::gui::{Button, Dialog};
 
 pub struct WindowsButton;
 
@@ -10,5 +10,14 @@ impl Button for WindowsButton {
 
     fn on_click(&self) {
         println!("Click! Hello, Windows!");
+    }
+}
+
+pub struct WindowsDialog;
+
+impl Dialog for WindowsDialog {
+    /// Creates a Windows button.
+    fn create_button(&self) -> Box<dyn Button> {
+        Box::new(WindowsButton)
     }
 }
